Minor style and English grammar suggestions
Friendly peer review of the first section of the README.

Only one backtick required for code spans.

Added link to meaning of Xanadu

Added link to meaning of Xanadu

The link is for those people (not **cough** me) that do not immediately recognize the significance of aspiring to be the "CLI Xanadu".