Minimize lodash load
This PR replaces the use of the monolithic `lodash` dependence with the much lighter node modules that offer selective functionality.

It also removes lodash from packages/core/package.json. I don't see any use of it there, so it seems that we can elmininate it entirely from that package.json?