feat: Add prefill option to input prompt
This update introduces a new `prefill` option to the input prompt.
This governs the behaviour of how the default value (if provided) is presented and
edited.

The `prefill` option supports two modes:
- `'tab'` (default): Pressing `tab` inlines the default value for editing, and
pressing `backspace` clears it.
- `'editable'`: The default value is immediately inlined and editable without
needing to press `tab`.

The README has been updated to reflect the new option.
Tests have been updated to verify the new behavior and provde that the change is BWC.

Closes issue: #1737